Cleaning up the merge from rails_2_maintenance
The move of three packages to net.sf.rails.game.financials is not
reflected on the maintenance branch.
