+++ src/main/java/net/sf/rails/game/financial/PlayerShareUtils.java
@@ -92,10 +92,15 @@ public class PlayerShareUtils {
         Player potential = company.findPlayerToDump();
         int potentialShareNumber = potential.getPortfolioModel().getShare(company);
         int shareNumberDumpDifference = presidentShareNumber - potentialShareNumber;
+        boolean presidentShareOnly = false;
+        
+        if (presidentCert.getShare() == presidentShareNumber)  { // Only President Share to be sold...
+            presidentShareOnly = true;
+        }
         
         // ... if this is less than what the pool allows => goes back to non-president selling
         int poolAllows = poolAllowsShareNumbers(company);
-        if (shareNumberDumpDifference <= poolAllows) {
+        if ((shareNumberDumpDifference <= poolAllows) && (!presidentShareOnly)) {
             return otherSellMultiple(company, president);
         }
         
@@ -128,7 +133,7 @@ public class PlayerShareUtils {
                     // d is the amount sold in addition to standard shares, returned has the remaining part of the president share
                     int remaining = presidentCert.getShares() - d;
                     if (returnShareNumbers.contains(remaining)) {
-                        sharesToSell.add(s);
+                        sharesToSell.add(s+d);
                     }
                 } else {
                     break; // pool is full
@@ -177,6 +182,13 @@ public class PlayerShareUtils {
                 nbCertsToSell--;
                 if (nbCertsToSell == 0) {
                     break;
+                }
+            }
+                else if (cert.isPresidentShare() && cert.getShares()== shareUnits) {
+                    certsToSell.add(cert);
+                    nbCertsToSell--;
+                    if (nbCertsToSell == 0) {
+                        break;
                 }
             }
         }
